Fix file name and add basic tests for class.
